fs: Move fsp_srv subclasses to separate files
fs: Move additional files to the fsp directory
