+++ src/core/hle/service/filesystem/fsp/fsp_ldr.cpp
@@ -1,23 +1,22 @@
 // SPDX-FileCopyrightText: Copyright 2018 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "core/hle/service/filesystem/fsp_pr.h"
+#include "core/hle/service/filesystem/fsp/fsp_ldr.h"
 
 namespace Service::FileSystem {
 
-FSP_PR::FSP_PR(Core::System& system_) : ServiceFramework{system_, "fsp:pr"} {
+FSP_LDR::FSP_LDR(Core::System& system_) : ServiceFramework{system_, "fsp:ldr"} {
     // clang-format off
     static const FunctionInfo functions[] = {
-        {0, nullptr, "RegisterProgram"},
-        {1, nullptr, "UnregisterProgram"},
+        {0, nullptr, "OpenCodeFileSystem"},
+        {1, nullptr, "IsArchivedProgram"},
         {2, nullptr, "SetCurrentProcess"},
-        {256, nullptr, "SetEnabledProgramVerification"},
     };
     // clang-format on
 
     RegisterHandlers(functions);
 }
 
-FSP_PR::~FSP_PR() = default;
+FSP_LDR::~FSP_LDR() = default;
 
 } // namespace Service::FileSystem
