+++ src/core/hle/service/filesystem/fsp/fsp_srv.cpp
@@ -26,7 +26,9 @@
 #include "core/file_sys/vfs.h"
 #include "core/hle/result.h"
 #include "core/hle/service/filesystem/filesystem.h"
-#include "core/hle/service/filesystem/fsp_srv.h"
+#include "core/hle/service/filesystem/fsp/fs_i_filesystem.h"
+#include "core/hle/service/filesystem/fsp/fs_i_storage.h"
+#include "core/hle/service/filesystem/fsp/fsp_srv.h"
 #include "core/hle/service/filesystem/romfs_controller.h"
 #include "core/hle/service/filesystem/save_data_controller.h"
 #include "core/hle/service/hle_ipc.h"
@@ -34,19 +36,6 @@
 #include "core/reporter.h"
 
 namespace Service::FileSystem {
-
-struct SizeGetter {
-    std::function<u64()> get_free_size;
-    std::function<u64()> get_total_size;
-
-    static SizeGetter FromStorageId(const FileSystemController& fsc, FileSys::StorageId id) {
-        return {
-            [&fsc, id] { return fsc.GetFreeSpaceSize(id); },
-            [&fsc, id] { return fsc.GetTotalSpaceSize(id); },
-        };
-    }
-};
-
 enum class FileSystemType : u8 {
     Invalid0 = 0,
     Invalid1 = 1,
@@ -58,532 +47,13 @@ enum class FileSystemType : u8 {
     ApplicationPackage = 7,
 };
 
-class IStorage final : public ServiceFramework<IStorage> {
-public:
-    explicit IStorage(Core::System& system_, FileSys::VirtualFile backend_)
-        : ServiceFramework{system_, "IStorage"}, backend(std::move(backend_)) {
-        static const FunctionInfo functions[] = {
-            {0, &IStorage::Read, "Read"},
-            {1, nullptr, "Write"},
-            {2, nullptr, "Flush"},
-            {3, nullptr, "SetSize"},
-            {4, &IStorage::GetSize, "GetSize"},
-            {5, nullptr, "OperateRange"},
-        };
-        RegisterHandlers(functions);
-    }
-
-private:
-    FileSys::VirtualFile backend;
-
-    void Read(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-        const s64 offset = rp.Pop<s64>();
-        const s64 length = rp.Pop<s64>();
-
-        LOG_DEBUG(Service_FS, "called, offset=0x{:X}, length={}", offset, length);
-
-        // Error checking
-        if (length < 0) {
-            LOG_ERROR(Service_FS, "Length is less than 0, length={}", length);
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(FileSys::ERROR_INVALID_SIZE);
-            return;
-        }
-        if (offset < 0) {
-            LOG_ERROR(Service_FS, "Offset is less than 0, offset={}", offset);
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(FileSys::ERROR_INVALID_OFFSET);
-            return;
-        }
-
-        // Read the data from the Storage backend
-        std::vector<u8> output = backend->ReadBytes(length, offset);
-        // Write the data to memory
-        ctx.WriteBuffer(output);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(ResultSuccess);
-    }
-
-    void GetSize(HLERequestContext& ctx) {
-        const u64 size = backend->GetSize();
-        LOG_DEBUG(Service_FS, "called, size={}", size);
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push<u64>(size);
-    }
-};
-
-class IFile final : public ServiceFramework<IFile> {
-public:
-    explicit IFile(Core::System& system_, FileSys::VirtualFile backend_)
-        : ServiceFramework{system_, "IFile"}, backend(std::move(backend_)) {
-        static const FunctionInfo functions[] = {
-            {0, &IFile::Read, "Read"},
-            {1, &IFile::Write, "Write"},
-            {2, &IFile::Flush, "Flush"},
-            {3, &IFile::SetSize, "SetSize"},
-            {4, &IFile::GetSize, "GetSize"},
-            {5, nullptr, "OperateRange"},
-            {6, nullptr, "OperateRangeWithBuffer"},
-        };
-        RegisterHandlers(functions);
-    }
-
-private:
-    FileSys::VirtualFile backend;
-
-    void Read(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-        const u64 option = rp.Pop<u64>();
-        const s64 offset = rp.Pop<s64>();
-        const s64 length = rp.Pop<s64>();
-
-        LOG_DEBUG(Service_FS, "called, option={}, offset=0x{:X}, length={}", option, offset,
-                  length);
-
-        // Error checking
-        if (length < 0) {
-            LOG_ERROR(Service_FS, "Length is less than 0, length={}", length);
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(FileSys::ERROR_INVALID_SIZE);
-            return;
-        }
-        if (offset < 0) {
-            LOG_ERROR(Service_FS, "Offset is less than 0, offset={}", offset);
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(FileSys::ERROR_INVALID_OFFSET);
-            return;
-        }
-
-        // Read the data from the Storage backend
-        std::vector<u8> output = backend->ReadBytes(length, offset);
-
-        // Write the data to memory
-        ctx.WriteBuffer(output);
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push(static_cast<u64>(output.size()));
-    }
-
-    void Write(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-        const u64 option = rp.Pop<u64>();
-        const s64 offset = rp.Pop<s64>();
-        const s64 length = rp.Pop<s64>();
-
-        LOG_DEBUG(Service_FS, "called, option={}, offset=0x{:X}, length={}", option, offset,
-                  length);
-
-        // Error checking
-        if (length < 0) {
-            LOG_ERROR(Service_FS, "Length is less than 0, length={}", length);
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(FileSys::ERROR_INVALID_SIZE);
-            return;
-        }
-        if (offset < 0) {
-            LOG_ERROR(Service_FS, "Offset is less than 0, offset={}", offset);
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(FileSys::ERROR_INVALID_OFFSET);
-            return;
-        }
-
-        const auto data = ctx.ReadBuffer();
-
-        ASSERT_MSG(
-            static_cast<s64>(data.size()) <= length,
-            "Attempting to write more data than requested (requested={:016X}, actual={:016X}).",
-            length, data.size());
-
-        // Write the data to the Storage backend
-        const auto write_size =
-            static_cast<std::size_t>(std::distance(data.begin(), data.begin() + length));
-        const std::size_t written = backend->Write(data.data(), write_size, offset);
-
-        ASSERT_MSG(static_cast<s64>(written) == length,
-                   "Could not write all bytes to file (requested={:016X}, actual={:016X}).", length,
-                   written);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(ResultSuccess);
-    }
-
-    void Flush(HLERequestContext& ctx) {
-        LOG_DEBUG(Service_FS, "called");
-
-        // Exists for SDK compatibiltity -- No need to flush file.
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(ResultSuccess);
-    }
-
-    void SetSize(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-        const u64 size = rp.Pop<u64>();
-        LOG_DEBUG(Service_FS, "called, size={}", size);
-
-        backend->Resize(size);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(ResultSuccess);
-    }
-
-    void GetSize(HLERequestContext& ctx) {
-        const u64 size = backend->GetSize();
-        LOG_DEBUG(Service_FS, "called, size={}", size);
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push<u64>(size);
-    }
-};
-
-template <typename T>
-static void BuildEntryIndex(std::vector<FileSys::Entry>& entries, const std::vector<T>& new_data,
-                            FileSys::EntryType type) {
-    entries.reserve(entries.size() + new_data.size());
-
-    for (const auto& new_entry : new_data) {
-        auto name = new_entry->GetName();
-
-        if (type == FileSys::EntryType::File && name == FileSys::GetSaveDataSizeFileName()) {
-            continue;
-        }
-
-        entries.emplace_back(name, type,
-                             type == FileSys::EntryType::Directory ? 0 : new_entry->GetSize());
-    }
-}
-
-class IDirectory final : public ServiceFramework<IDirectory> {
-public:
-    explicit IDirectory(Core::System& system_, FileSys::VirtualDir backend_, OpenDirectoryMode mode)
-        : ServiceFramework{system_, "IDirectory"}, backend(std::move(backend_)) {
-        static const FunctionInfo functions[] = {
-            {0, &IDirectory::Read, "Read"},
-            {1, &IDirectory::GetEntryCount, "GetEntryCount"},
-        };
-        RegisterHandlers(functions);
-
-        // TODO(DarkLordZach): Verify that this is the correct behavior.
-        // Build entry index now to save time later.
-        if (True(mode & OpenDirectoryMode::Directory)) {
-            BuildEntryIndex(entries, backend->GetSubdirectories(), FileSys::EntryType::Directory);
-        }
-        if (True(mode & OpenDirectoryMode::File)) {
-            BuildEntryIndex(entries, backend->GetFiles(), FileSys::EntryType::File);
-        }
-    }
-
-private:
-    FileSys::VirtualDir backend;
-    std::vector<FileSys::Entry> entries;
-    u64 next_entry_index = 0;
-
-    void Read(HLERequestContext& ctx) {
-        LOG_DEBUG(Service_FS, "called.");
-
-        // Calculate how many entries we can fit in the output buffer
-        const u64 count_entries = ctx.GetWriteBufferNumElements<FileSys::Entry>();
-
-        // Cap at total number of entries.
-        const u64 actual_entries = std::min(count_entries, entries.size() - next_entry_index);
-
-        // Determine data start and end
-        const auto* begin = reinterpret_cast<u8*>(entries.data() + next_entry_index);
-        const auto* end = reinterpret_cast<u8*>(entries.data() + next_entry_index + actual_entries);
-        const auto range_size = static_cast<std::size_t>(std::distance(begin, end));
-
-        next_entry_index += actual_entries;
-
-        // Write the data to memory
-        ctx.WriteBuffer(begin, range_size);
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push(actual_entries);
-    }
-
-    void GetEntryCount(HLERequestContext& ctx) {
-        LOG_DEBUG(Service_FS, "called");
-
-        u64 count = entries.size() - next_entry_index;
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push(count);
-    }
-};
-
-class IFileSystem final : public ServiceFramework<IFileSystem> {
-public:
-    explicit IFileSystem(Core::System& system_, FileSys::VirtualDir backend_, SizeGetter size_)
-        : ServiceFramework{system_, "IFileSystem"}, backend{std::move(backend_)}, size{std::move(
-                                                                                      size_)} {
-        static const FunctionInfo functions[] = {
-            {0, &IFileSystem::CreateFile, "CreateFile"},
-            {1, &IFileSystem::DeleteFile, "DeleteFile"},
-            {2, &IFileSystem::CreateDirectory, "CreateDirectory"},
-            {3, &IFileSystem::DeleteDirectory, "DeleteDirectory"},
-            {4, &IFileSystem::DeleteDirectoryRecursively, "DeleteDirectoryRecursively"},
-            {5, &IFileSystem::RenameFile, "RenameFile"},
-            {6, nullptr, "RenameDirectory"},
-            {7, &IFileSystem::GetEntryType, "GetEntryType"},
-            {8, &IFileSystem::OpenFile, "OpenFile"},
-            {9, &IFileSystem::OpenDirectory, "OpenDirectory"},
-            {10, &IFileSystem::Commit, "Commit"},
-            {11, &IFileSystem::GetFreeSpaceSize, "GetFreeSpaceSize"},
-            {12, &IFileSystem::GetTotalSpaceSize, "GetTotalSpaceSize"},
-            {13, &IFileSystem::CleanDirectoryRecursively, "CleanDirectoryRecursively"},
-            {14, &IFileSystem::GetFileTimeStampRaw, "GetFileTimeStampRaw"},
-            {15, nullptr, "QueryEntry"},
-            {16, &IFileSystem::GetFileSystemAttribute, "GetFileSystemAttribute"},
-        };
-        RegisterHandlers(functions);
-    }
-
-    void CreateFile(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        const u64 file_mode = rp.Pop<u64>();
-        const u32 file_size = rp.Pop<u32>();
-
-        LOG_DEBUG(Service_FS, "called. file={}, mode=0x{:X}, size=0x{:08X}", name, file_mode,
-                  file_size);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.CreateFile(name, file_size));
-    }
-
-    void DeleteFile(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_DEBUG(Service_FS, "called. file={}", name);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.DeleteFile(name));
-    }
-
-    void CreateDirectory(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_DEBUG(Service_FS, "called. directory={}", name);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.CreateDirectory(name));
-    }
-
-    void DeleteDirectory(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_DEBUG(Service_FS, "called. directory={}", name);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.DeleteDirectory(name));
-    }
-
-    void DeleteDirectoryRecursively(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_DEBUG(Service_FS, "called. directory={}", name);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.DeleteDirectoryRecursively(name));
-    }
-
-    void CleanDirectoryRecursively(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_DEBUG(Service_FS, "called. Directory: {}", name);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.CleanDirectoryRecursively(name));
-    }
-
-    void RenameFile(HLERequestContext& ctx) {
-        const std::string src_name = Common::StringFromBuffer(ctx.ReadBuffer(0));
-        const std::string dst_name = Common::StringFromBuffer(ctx.ReadBuffer(1));
-
-        LOG_DEBUG(Service_FS, "called. file '{}' to file '{}'", src_name, dst_name);
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(backend.RenameFile(src_name, dst_name));
-    }
-
-    void OpenFile(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        const auto mode = static_cast<FileSys::Mode>(rp.Pop<u32>());
-
-        LOG_DEBUG(Service_FS, "called. file={}, mode={}", name, mode);
-
-        FileSys::VirtualFile vfs_file{};
-        auto result = backend.OpenFile(&vfs_file, name, mode);
-        if (result != ResultSuccess) {
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(result);
-            return;
-        }
-
-        auto file = std::make_shared<IFile>(system, vfs_file);
-
-        IPC::ResponseBuilder rb{ctx, 2, 0, 1};
-        rb.Push(ResultSuccess);
-        rb.PushIpcInterface<IFile>(std::move(file));
-    }
-
-    void OpenDirectory(HLERequestContext& ctx) {
-        IPC::RequestParser rp{ctx};
-
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-        const auto mode = rp.PopRaw<OpenDirectoryMode>();
-
-        LOG_DEBUG(Service_FS, "called. directory={}, mode={}", name, mode);
-
-        FileSys::VirtualDir vfs_dir{};
-        auto result = backend.OpenDirectory(&vfs_dir, name);
-        if (result != ResultSuccess) {
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(result);
-            return;
-        }
-
-        auto directory = std::make_shared<IDirectory>(system, vfs_dir, mode);
-
-        IPC::ResponseBuilder rb{ctx, 2, 0, 1};
-        rb.Push(ResultSuccess);
-        rb.PushIpcInterface<IDirectory>(std::move(directory));
-    }
-
-    void GetEntryType(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_DEBUG(Service_FS, "called. file={}", name);
-
-        FileSys::EntryType vfs_entry_type{};
-        auto result = backend.GetEntryType(&vfs_entry_type, name);
-        if (result != ResultSuccess) {
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(result);
-            return;
-        }
-
-        IPC::ResponseBuilder rb{ctx, 3};
-        rb.Push(ResultSuccess);
-        rb.Push<u32>(static_cast<u32>(vfs_entry_type));
-    }
-
-    void Commit(HLERequestContext& ctx) {
-        LOG_WARNING(Service_FS, "(STUBBED) called");
-
-        IPC::ResponseBuilder rb{ctx, 2};
-        rb.Push(ResultSuccess);
-    }
-
-    void GetFreeSpaceSize(HLERequestContext& ctx) {
-        LOG_DEBUG(Service_FS, "called");
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push(size.get_free_size());
-    }
-
-    void GetTotalSpaceSize(HLERequestContext& ctx) {
-        LOG_DEBUG(Service_FS, "called");
-
-        IPC::ResponseBuilder rb{ctx, 4};
-        rb.Push(ResultSuccess);
-        rb.Push(size.get_total_size());
-    }
-
-    void GetFileTimeStampRaw(HLERequestContext& ctx) {
-        const auto file_buffer = ctx.ReadBuffer();
-        const std::string name = Common::StringFromBuffer(file_buffer);
-
-        LOG_WARNING(Service_FS, "(Partial Implementation) called. file={}", name);
-
-        FileSys::FileTimeStampRaw vfs_timestamp{};
-        auto result = backend.GetFileTimeStampRaw(&vfs_timestamp, name);
-        if (result != ResultSuccess) {
-            IPC::ResponseBuilder rb{ctx, 2};
-            rb.Push(result);
-            return;
-        }
-
-        IPC::ResponseBuilder rb{ctx, 10};
-        rb.Push(ResultSuccess);
-        rb.PushRaw(vfs_timestamp);
-    }
-
-    void GetFileSystemAttribute(HLERequestContext& ctx) {
-        LOG_WARNING(Service_FS, "(STUBBED) called");
-
-        struct FileSystemAttribute {
-            u8 dir_entry_name_length_max_defined;
-            u8 file_entry_name_length_max_defined;
-            u8 dir_path_name_length_max_defined;
-            u8 file_path_name_length_max_defined;
-            INSERT_PADDING_BYTES_NOINIT(0x5);
-            u8 utf16_dir_entry_name_length_max_defined;
-            u8 utf16_file_entry_name_length_max_defined;
-            u8 utf16_dir_path_name_length_max_defined;
-            u8 utf16_file_path_name_length_max_defined;
-            INSERT_PADDING_BYTES_NOINIT(0x18);
-            s32 dir_entry_name_length_max;
-            s32 file_entry_name_length_max;
-            s32 dir_path_name_length_max;
-            s32 file_path_name_length_max;
-            INSERT_PADDING_WORDS_NOINIT(0x5);
-            s32 utf16_dir_entry_name_length_max;
-            s32 utf16_file_entry_name_length_max;
-            s32 utf16_dir_path_name_length_max;
-            s32 utf16_file_path_name_length_max;
-            INSERT_PADDING_WORDS_NOINIT(0x18);
-            INSERT_PADDING_WORDS_NOINIT(0x1);
-        };
-        static_assert(sizeof(FileSystemAttribute) == 0xc0,
-                      "FileSystemAttribute has incorrect size");
-
-        FileSystemAttribute savedata_attribute{};
-        savedata_attribute.dir_entry_name_length_max_defined = true;
-        savedata_attribute.file_entry_name_length_max_defined = true;
-        savedata_attribute.dir_entry_name_length_max = 0x40;
-        savedata_attribute.file_entry_name_length_max = 0x40;
-
-        IPC::ResponseBuilder rb{ctx, 50};
-        rb.Push(ResultSuccess);
-        rb.PushRaw(savedata_attribute);
-    }
-
-private:
-    VfsDirectoryServiceWrapper backend;
-    SizeGetter size;
-};
-
 class ISaveDataInfoReader final : public ServiceFramework<ISaveDataInfoReader> {
 public:
     explicit ISaveDataInfoReader(Core::System& system_,
                                  std::shared_ptr<SaveDataController> save_data_controller_,
                                  FileSys::SaveDataSpaceId space)
-        : ServiceFramework{system_, "ISaveDataInfoReader"}, save_data_controller{
-                                                                save_data_controller_} {
+        : ServiceFramework{system_, "ISaveDataInfoReader"},
+          save_data_controller{save_data_controller_} {
         static const FunctionInfo functions[] = {
             {0, &ISaveDataInfoReader::ReadSaveDataInfo, "ReadSaveDataInfo"},
         };
